fix parse error

+++ backend/src/test/java/com/bakdata/conquery/models/events/stores/types/ColumnStoreSerializationTests.java
@@ -109,7 +109,7 @@ public class ColumnStoreSerializationTests {
 				new EncodedStringStore(new DictionaryStore(IntArrayStore.create(10), DICTIONARY), Encoding.Base16LowerCase),
 				new PrefixSuffixStringStore(new EncodedStringStore(new DictionaryStore(IntArrayStore.create(10), DICTIONARY), Encoding.Base16LowerCase), "a", "b"),
 
-				new NumberStringStore(new Range.LongRange(0L, 7L), ByteArrayStore.create(10)),
+				new NumberStringStore(new Range.IntegerRange(0, 7), ByteArrayStore.create(10)),
 				new SingletonStringStore("a", BitSetStore.create(10)),
 				new DirectDateRangeStore(IntegerDateStore.create(10), IntegerDateStore.create(10)),
 				new QuarterDateRangeStore(LongArrayStore.create(10)),
